refactor: rename `base` => `_base` and remove a redundant method
